EZP-28038: Refined `ezcountry` fieldtype
